backend/requirements.txt
# SureCare AI — Production Dependencies

# ── Core Framework ──────────────────────────────────────────
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# ── Database (Supabase PostgreSQL) ──────────────────────────
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# ── AI / LLM ───────────────────────────────────────────────
google-generativeai==0.8.3
numpy==1.26.4

# ── Document Processing & OCR ──────────────────────────────
pypdf==4.1.0
pytesseract==0.3.13
Pillow==10.2.0
python-docx==1.1.0

# ── HTTP Client (Supabase Storage REST) ────────────────────
httpx==0.27.0
aiofiles==23.2.1

# ── Security ────────────────────────────────────────────────
bcrypt==4.1.2

# ── FAISS & Embeddings (optional) ──────────────────────────
faiss-cpu==1.9.0.post1
sentence-transformers==2.7.0
